Trait moka::sync::ConcurrentCacheExt

pub trait ConcurrentCacheExt<K, V> {
    fn sync(&self);
}

Provides extra methods that will be useful for testing.

Required methods

fn sync(&self)

Performs any pending maintenance operations needed by the cache.