moka::sync

Struct Cache

pub struct Cache<K, V, S = RandomState> { /* private fields */ }

A thread-safe concurrent synchronous in-memory cache.

Cache supports full concurrency of retrievals and a high expected concurrency for updates.

Cache utilizes a lock-free concurrent hash table as the central key-value storage. Cache performs a best-effort bounding of the map using an entry replacement algorithm to determine which entries to evict when the capacity is exceeded.

Example: insert, get and invalidate

Avoiding to clone the value at get

Sharing a cache across threads

No lock is needed

Hashing Algorithm

Example: Size-based Eviction

Example: Time-based Expirations

Cache-level TTL and TTI policies

Per-entry expiration policy

Example: Eviction Listener

You should avoid eviction listener to panic

Example: insert, get and invalidate

Cache entries are manually added using insert or get_with methods, and are stored in the cache until either evicted or manually invalidated.

Here’s an example of reading and updating a cache by using multiple threads:

use moka::sync::Cache;

use std::thread;

fn value(n: usize) -> String {
    format!("value {n}")
}

const NUM_THREADS: usize = 16;
const NUM_KEYS_PER_THREAD: usize = 64;

// Create a cache that can store up to 10,000 entries.
let cache = Cache::new(10_000);

// Spawn threads and read and update the cache simultaneously.
let threads: Vec<_> = (0..NUM_THREADS)
    .map(|i| {
        // To share the same cache across the threads, clone it.
        // This is a cheap operation.
        let my_cache = cache.clone();
        let start = i * NUM_KEYS_PER_THREAD;
        let end = (i + 1) * NUM_KEYS_PER_THREAD;

        thread::spawn(move || {
            // Insert 64 entries. (NUM_KEYS_PER_THREAD = 64)
            for key in start..end {
                my_cache.insert(key, value(key));
                // get() returns Option<String>, a clone of the stored value.
                assert_eq!(my_cache.get(&key), Some(value(key)));
            }

            // Invalidate every 4 element of the inserted entries.
            for key in (start..end).step_by(4) {
                my_cache.invalidate(&key);
            }
        })
    })
    .collect();

// Wait for all threads to complete.
threads.into_iter().for_each(|t| t.join().expect("Failed"));

// Verify the result.
for key in 0..(NUM_THREADS * NUM_KEYS_PER_THREAD) {
    if key % 4 == 0 {
        assert_eq!(cache.get(&key), None);
    } else {
        assert_eq!(cache.get(&key), Some(value(key)));
    }
}

If you want to atomically initialize and insert a value when the key is not present, you might want to check other insertion methods get_with and try_get_with.

Avoiding to clone the value at get

The return type of get method is Option<V> instead of Option<&V>. Every time get is called for an existing key, it creates a clone of the stored value V and returns it. This is because the Cache allows concurrent updates from threads so a value stored in the cache can be dropped or replaced at any time by any other thread. get cannot return a reference &V as it is impossible to guarantee the value outlives the reference.

If you want to store values that will be expensive to clone, wrap them by std::sync::Arc before storing in a cache. Arc is a thread-safe reference-counted pointer and its clone() method is cheap.

Sharing a cache across threads

To share a cache across threads, do one of the followings:

Create a clone of the cache by calling its clone method and pass it to other thread.

Wrap the cache by a sync::OnceCell or sync::Lazy from once_cell create, and set it to a static variable.

Cloning is a cheap operation for Cache as it only creates thread-safe reference-counted pointers to the internal data structures.

No lock is needed

Don’t wrap a Cache by a lock such as Mutex or RwLock. All methods provided by the Cache are considered thread-safe, and can be safely called by multiple threads at the same time. No lock is needed.

Hashing Algorithm

By default, Cache uses a hashing algorithm selected to provide resistance against HashDoS attacks. It will be the same one used by std::collections::HashMap, which is currently SipHash 1-3.

While SipHash’s performance is very competitive for medium sized keys, other hashing algorithms will outperform it for small keys such as integers as well as large keys such as long strings. However those algorithms will typically not protect against attacks such as HashDoS.

The hashing algorithm can be replaced on a per-Cache basis using the build_with_hasher method of the CacheBuilder. Many alternative algorithms are available on crates.io, such as the AHash crate.

Example: Size-based Eviction

use moka::sync::Cache;

// Evict based on the number of entries in the cache.
let cache = Cache::builder()
    // Up to 10,000 entries.
    .max_capacity(10_000)
    // Create the cache.
    .build();
cache.insert(1, "one".to_string());

// Evict based on the byte length of strings in the cache.
let cache = Cache::builder()
    // A weigher closure takes &K and &V and returns a u32
    // representing the relative size of the entry.
    .weigher(|_key, value: &String| -> u32 {
        value.len().try_into().unwrap_or(u32::MAX)
    })
    // This cache will hold up to 32MiB of values.
    .max_capacity(32 * 1024 * 1024)
    .build();
cache.insert(2, "two".to_string());

If your cache should not grow beyond a certain size, use the max_capacity method of the CacheBuilder to set the upper bound. The cache will try to evict entries that have not been used recently or very often.

At the cache creation time, a weigher closure can be set by the weigher method of the CacheBuilder. A weigher closure takes &K and &V as the arguments and returns a u32 representing the relative size of the entry:

If the weigher is not set, the cache will treat each entry has the same size of 1. This means the cache will be bounded by the number of entries.

If the weigher is set, the cache will call the weigher to calculate the weighted size (relative size) on an entry. This means the cache will be bounded by the total weighted size of entries.

Note that weighted sizes are not used when making eviction selections.

Example: Time-based Expirations

Cache-level TTL and TTI policies

Cache supports the following cache-level expiration policies:

Time to live (TTL): A cached entry will be expired after the specified duration past from insert.

Time to idle (TTI): A cached entry will be expired after the specified duration past from get or insert.

They are a cache-level expiration policies; all entries in the cache will have the same TTL and/or TTI durations. If you want to set different expiration durations for different entries, see the next section.

use moka::sync::Cache;
use std::time::Duration;

let cache = Cache::builder()
    // Time to live (TTL): 30 minutes
    .time_to_live(Duration::from_secs(30 * 60))
    // Time to idle (TTI):  5 minutes
    .time_to_idle(Duration::from_secs( 5 * 60))
    // Create the cache.
    .build();

// This entry will expire after 5 minutes (TTI) if there is no get().
cache.insert(0, "zero");

// This get() will extend the entry life for another 5 minutes.
cache.get(&0);

// Even though we keep calling get(), the entry will expire
// after 30 minutes (TTL) from the insert().

Per-entry expiration policy

Cache supports per-entry expiration policy through the Expiry trait.

Expiry trait provides three callback methods: expire_after_create, expire_after_read and expire_after_update. When a cache entry is inserted, read or updated, one of these methods is called. These methods return an Option<Duration>, which is used as the expiration duration of the entry.

Expiry trait provides the default implementations of these methods, so you will implement only the methods you want to customize.

use moka::{sync::Cache, Expiry};
use std::time::{Duration, Instant};

// In this example, we will create a `sync::Cache` with `u32` as the key, and
// `(Expiration, String)` as the value. `Expiration` is an enum to represent the
// expiration of the value, and `String` is the application data of the value.

/// An enum to represent the expiration of a value.
#[derive(Clone, Copy, Debug, Eq, PartialEq)]
pub enum Expiration {
    /// The value never expires.
    Never,
    /// The value expires after a short time. (5 seconds in this example)
    AfterShortTime,
    /// The value expires after a long time. (15 seconds in this example)
    AfterLongTime,
}

impl Expiration {
    /// Returns the duration of this expiration.
    pub fn as_duration(&self) -> Option<Duration> {
        match self {
            Expiration::Never => None,
            Expiration::AfterShortTime => Some(Duration::from_secs(5)),
            Expiration::AfterLongTime => Some(Duration::from_secs(15)),
        }
    }
}

/// An expiry that implements `moka::Expiry` trait. `Expiry` trait provides the
/// default implementations of three callback methods `expire_after_create`,
/// `expire_after_read`, and `expire_after_update`.
///
/// In this example, we only override the `expire_after_create` method.
pub struct MyExpiry;

impl Expiry<u32, (Expiration, String)> for MyExpiry {
    /// Returns the duration of the expiration of the value that was just
    /// created.
    fn expire_after_create(
        &self,
        _key: &u32,
        value: &(Expiration, String),
        _current_time: Instant,
    ) -> Option<Duration> {
        let duration = value.0.as_duration();
        println!("MyExpiry: expire_after_create called with key {_key} and value {value:?}. Returning {duration:?}.");
        duration
    }
}

// Create a `Cache<u32, (Expiration, String)>` with an expiry `MyExpiry` and
// eviction listener.
let expiry = MyExpiry;

let eviction_listener = |key, _value, cause| {
    println!("Evicted key {key}. Cause: {cause:?}");
};

let cache = Cache::builder()
    .max_capacity(100)
    .expire_after(expiry)
    .eviction_listener(eviction_listener)
    .build();

// Insert some entries into the cache with different expirations.
cache.get_with(0, || (Expiration::AfterShortTime, "a".to_string()));
cache.get_with(1, || (Expiration::AfterLongTime, "b".to_string()));
cache.get_with(2, || (Expiration::Never, "c".to_string()));

// Verify that all the inserted entries exist.
assert!(cache.contains_key(&0));
assert!(cache.contains_key(&1));
assert!(cache.contains_key(&2));

// Sleep for 6 seconds. Key 0 should expire.
println!("\nSleeping for 6 seconds...\n");
std::thread::sleep(Duration::from_secs(6));
println!("Entry count: {}", cache.entry_count());

// Verify that key 0 has been evicted.
assert!(!cache.contains_key(&0));
assert!(cache.contains_key(&1));
assert!(cache.contains_key(&2));

// Sleep for 10 more seconds. Key 1 should expire.
println!("\nSleeping for 10 seconds...\n");
std::thread::sleep(Duration::from_secs(10));
println!("Entry count: {}", cache.entry_count());

// Verify that key 1 has been evicted.
assert!(!cache.contains_key(&1));
assert!(cache.contains_key(&2));

// Manually invalidate key 2.
cache.invalidate(&2);
assert!(!cache.contains_key(&2));

println!("\nSleeping for a second...\n");
std::thread::sleep(Duration::from_secs(1));
println!("Entry count: {}", cache.entry_count());

println!("\nDone!");

Example: Eviction Listener

A Cache can be configured with an eviction listener, a closure that is called every time there is a cache eviction. The listener takes three parameters: the key and value of the evicted entry, and the RemovalCause to indicate why the entry was evicted.

An eviction listener can be used to keep other data structures in sync with the cache, for example.

The following example demonstrates how to use an eviction listener with time-to-live expiration to manage the lifecycle of temporary files on a filesystem. The cache stores the paths of the files, and when one of them has expired, the eviction listener will be called with the path, so it can remove the file from the filesystem.

// Cargo.toml
//
// [dependencies]
// anyhow = "1.0"
// uuid = { version = "1.1", features = ["v4"] }

use moka::{sync::Cache, notification};

use anyhow::{anyhow, Context};
use std::{
    fs, io,
    path::{Path, PathBuf},
    sync::{Arc, RwLock},
    time::Duration,
};
use uuid::Uuid;

/// The DataFileManager writes, reads and removes data files.
struct DataFileManager {
    base_dir: PathBuf,
    file_count: usize,
}

impl DataFileManager {
    fn new(base_dir: PathBuf) -> Self {
        Self {
            base_dir,
            file_count: 0,
        }
    }

    fn write_data_file(
        &mut self,
        key: impl AsRef<str>,
        contents: String
    ) -> io::Result<PathBuf> {
        // Use the key as a part of the filename.
        let mut path = self.base_dir.to_path_buf();
        path.push(key.as_ref());

        assert!(!path.exists(), "Path already exists: {path:?}");

        // create the file at the path and write the contents to the file.
        fs::write(&path, contents)?;
        self.file_count += 1;
        println!("Created a data file at {path:?} (file count: {})", self.file_count);
        Ok(path)
    }

    fn read_data_file(&self, path: impl AsRef<Path>) -> io::Result<String> {
        // Reads the contents of the file at the path, and return the contents.
        fs::read_to_string(path)
    }

    fn remove_data_file(&mut self, path: impl AsRef<Path>) -> io::Result<()> {
        // Remove the file at the path.
        fs::remove_file(path.as_ref())?;
        self.file_count -= 1;
        println!(
            "Removed a data file at {:?} (file count: {})",
            path.as_ref(),
            self.file_count
        );

        Ok(())
    }
}

fn main() -> anyhow::Result<()> {
    // Create an instance of the DataFileManager and wrap it with
    // Arc<RwLock<_>> so it can be shared across threads.
    let mut base_dir = std::env::temp_dir();
    base_dir.push(Uuid::new_v4().as_hyphenated().to_string());
    println!("base_dir: {base_dir:?}");
    std::fs::create_dir(&base_dir)?;

    let file_mgr = DataFileManager::new(base_dir);
    let file_mgr = Arc::new(RwLock::new(file_mgr));

    let file_mgr1 = Arc::clone(&file_mgr);

    // Create an eviction listener closure.
    let eviction_listener = move |k, v: PathBuf, cause| {
        // Try to remove the data file at the path `v`.
        println!("\n== An entry has been evicted. k: {k:?}, v: {v:?}, cause: {cause:?}");

        // Acquire the write lock of the DataFileManager. We must handle
        // error cases here to prevent the listener from panicking.
        match file_mgr1.write() {
            Err(_e) => {
                eprintln!("The lock has been poisoned");
            }
            Ok(mut mgr) => {
                // Remove the data file using the DataFileManager.
                if let Err(_e) = mgr.remove_data_file(v.as_path()) {
                    eprintln!("Failed to remove a data file at {v:?}");
                }
            }
        }
    };

    // Create the cache. Set time to live for two seconds and set the
    // eviction listener.
    let cache = Cache::builder()
        .max_capacity(100)
        .time_to_live(Duration::from_secs(2))
        .eviction_listener(eviction_listener)
        .build();

    // Insert an entry to the cache.
    // This will create and write a data file for the key "user1", store the
    // path of the file to the cache, and return it.
    println!("== try_get_with()");
    let key = "user1";
    let path = cache
        .try_get_with(key, || -> anyhow::Result<_> {
            let mut mgr = file_mgr
                .write()
                .map_err(|_e| anyhow::anyhow!("The lock has been poisoned"))?;
            let path = mgr
                .write_data_file(key, "user data".into())
                .with_context(|| format!("Failed to create a data file"))?;
            Ok(path)
        })
        .map_err(|e| anyhow!("{e}"))?;

    // Read the data file at the path and print the contents.
    println!("\n== read_data_file()");
    {
        let mgr = file_mgr
            .read()
            .map_err(|_e| anyhow::anyhow!("The lock has been poisoned"))?;
        let contents = mgr
            .read_data_file(path.as_path())
            .with_context(|| format!("Failed to read data from {path:?}"))?;
        println!("contents: {contents}");
    }

    // Sleep for five seconds. While sleeping, the cache entry for key "user1"
    // will be expired and evicted, so the eviction listener will be called to
    // remove the file.
    std::thread::sleep(Duration::from_secs(5));

    cache.run_pending_tasks();

    Ok(())
}

You should avoid eviction listener to panic

It is very important to make an eviction listener closure not to panic. Otherwise, the cache will stop calling the listener after a panic. This is an intended behavior because the cache cannot know whether it is memory safe or not to call the panicked listener again.

When a listener panics, the cache will swallow the panic and disable the listener. If you want to know when a listener panics and the reason of the panic, you can enable an optional logging feature of Moka and check error-level logs.

To enable the logging, do the followings:

In Cargo.toml, add the crate feature logging for moka.

Set the logging level for moka to error or any lower levels (warn, info, …):

If you are using the env_logger crate, you can achieve this by setting RUST_LOG environment variable to moka=error.

If you have more than one caches, you may want to set a distinct name for each cache by using cache builder’s name method. The name will appear in the log.

Implementations

impl<K, V, S> Cache<K, V, S>

pub fn name(&self) -> Option<&str>

Returns cache’s name.

pub fn policy(&self) -> Policy

Returns a read-only cache policy of this cache.

At this time, cache policy cannot be modified after cache creation. A future version may support to modify it.

pub fn entry_count(&self) -> u64

Returns an approximate number of entries in this cache.

The value returned is an estimate; the actual count may differ if there are concurrent insertions or removals, or if some entries are pending removal due to expiration. This inaccuracy can be mitigated by performing a run_pending_tasks first.

Example

use moka::sync::Cache;

let cache = Cache::new(10);
cache.insert('n', "Netherland Dwarf");
cache.insert('l', "Lop Eared");
cache.insert('d', "Dutch");

// Ensure an entry exists.
assert!(cache.contains_key(&'n'));

// However, followings may print stale number zeros instead of threes.
println!("{}", cache.entry_count());   // -> 0
println!("{}", cache.weighted_size()); // -> 0

// To mitigate the inaccuracy, Call `run_pending_tasks` method to run
// pending internal tasks.
cache.run_pending_tasks();

// Followings will print the actual numbers.
println!("{}", cache.entry_count());   // -> 3
println!("{}", cache.weighted_size()); // -> 3

pub fn weighted_size(&self) -> u64

Returns an approximate total weighted size of entries in this cache.

The value returned is an estimate; the actual size may differ if there are concurrent insertions or removals, or if some entries are pending removal due to expiration. This inaccuracy can be mitigated by performing a run_pending_tasks first. See entry_count for a sample code.

impl<K, V> Cache<K, V, RandomState>
where K: Hash + Eq + Send + Sync + 'static, V: Clone + Send + Sync + 'static,

pub fn new(max_capacity: u64) -> Self

Constructs a new Cache<K, V> that will store up to the max_capacity.

To adjust various configuration knobs such as initial_capacity or time_to_live, use the CacheBuilder.

pub fn builder() -> CacheBuilder<K, V, Cache<K, V, RandomState>>

Returns a CacheBuilder, which can builds a Cache or SegmentedCache with various configuration knobs.

impl<K, V, S> Cache<K, V, S>
where K: Hash + Eq + Send + Sync + 'static, V: Clone + Send + Sync + 'static, S: BuildHasher + Clone + Send + Sync + 'static,

pub fn contains_key<Q>(&self, key: &Q) -> bool
where Q: Equivalent<K> + Hash + ?Sized,

Returns true if the cache contains a value for the key.

Unlike the get method, this method is not considered a cache read operation, so it does not update the historic popularity estimator or reset the idle timer for the key.

The key may be any borrowed form of the cache’s key type, but Hash and Eq on the borrowed form must match those for the key type.

pub fn get<Q>(&self, key: &Q) -> Option<V>
where Q: Equivalent<K> + Hash + ?Sized,

Returns a clone of the value corresponding to the key.

If you want to store values that will be expensive to clone, wrap them by std::sync::Arc before storing in a cache. Arc is a thread-safe reference-counted pointer and its clone() method is cheap.

The key may be any borrowed form of the cache’s key type, but Hash and Eq on the borrowed form must match those for the key type.

pub fn entry(&self, key: K) -> OwnedKeyEntrySelector<'_, K, V, S>
where K: Hash + Eq,

Takes a key K and returns an OwnedKeyEntrySelector that can be used to select or insert an entry.

Example

use moka::sync::Cache;

let cache: Cache<String, u32> = Cache::new(100);
let key = "key1".to_string();

let entry = cache.entry(key.clone()).or_insert(3);
assert!(entry.is_fresh());
assert_eq!(entry.key(), &key);
assert_eq!(entry.into_value(), 3);

let entry = cache.entry(key).or_insert(6);
// Not fresh because the value was already in the cache.
assert!(!entry.is_fresh());
assert_eq!(entry.into_value(), 3);

pub fn entry_by_ref<'a, Q>( &'a self, key: &'a Q, ) -> RefKeyEntrySelector<'a, K, Q, V, S>
where Q: Equivalent<K> + ToOwned<Owned = K> + Hash + ?Sized,

Takes a reference &Q of a key and returns an RefKeyEntrySelector that can be used to select or insert an entry.

Example

use moka::sync::Cache;

let cache: Cache<String, u32> = Cache::new(100);
let key = "key1".to_string();

let entry = cache.entry_by_ref(&key).or_insert(3);
assert!(entry.is_fresh());
assert_eq!(entry.key(), &key);
assert_eq!(entry.into_value(), 3);

let entry = cache.entry_by_ref(&key).or_insert(6);
// Not fresh because the value was already in the cache.
assert!(!entry.is_fresh());
assert_eq!(entry.into_value(), 3);

pub fn get_with(&self, key: K, init: impl FnOnce() -> V) -> V

Returns a clone of the value corresponding to the key. If the value does not exist, evaluates the init closure and inserts the output.

Concurrent calls on the same key

This method guarantees that concurrent calls on the same not-existing key are coalesced into one evaluation of the init closure. Only one of the calls evaluates its closure, and other calls wait for that closure to complete.

The following code snippet demonstrates this behavior:

use moka::sync::Cache;
use std::{sync::Arc, thread};

const TEN_MIB: usize = 10 * 1024 * 1024; // 10MiB
let cache = Cache::new(100);

// Spawn four threads.
let threads: Vec<_> = (0..4_u8)
    .map(|task_id| {
        let my_cache = cache.clone();
        thread::spawn(move || {
            println!("Thread {task_id} started.");

            // Try to insert and get the value for key1. Although all four
            // threads will call `get_with` at the same time, the `init` closure
            // must be evaluated only once.
            let value = my_cache.get_with("key1", || {
                println!("Thread {task_id} inserting a value.");
                Arc::new(vec![0u8; TEN_MIB])
            });

            // Ensure the value exists now.
            assert_eq!(value.len(), TEN_MIB);
            assert!(my_cache.get(&"key1").is_some());

            println!("Thread {task_id} got the value. (len: {})", value.len());
        })
    })
    .collect();

// Wait all threads to complete.
threads
    .into_iter()
    .for_each(|t| t.join().expect("Thread failed"));

Result

The init closure was called exactly once by thread 1.

Other threads were blocked until thread 1 inserted the value.

Thread 1 started.
Thread 0 started.
Thread 3 started.
Thread 2 started.
Thread 1 inserting a value.
Thread 2 got the value. (len: 10485760)
Thread 1 got the value. (len: 10485760)
Thread 0 got the value. (len: 10485760)
Thread 3 got the value. (len: 10485760)

Panics

This method panics when the init closure has panicked. When it happens, only the caller whose init closure panicked will get the panic (e.g. only thread 1 in the above sample). If there are other calls in progress (e.g. thread 0, 2 and 3 above), this method will restart and resolve one of the remaining init closure.

pub fn get_with_by_ref<Q>(&self, key: &Q, init: impl FnOnce() -> V) -> V
where Q: Equivalent<K> + ToOwned<Owned = K> + Hash + ?Sized,

Similar to get_with, but instead of passing an owned key, you can pass a reference to the key. If the key does not exist in the cache, the key will be cloned to create new entry in the cache.

pub fn get_with_if( &self, key: K, init: impl FnOnce() -> V, replace_if: impl FnMut(&V) -> bool, ) -> V

👎Deprecated since 0.10.0: Replaced with entry().or_insert_with_if()

TODO: Remove this in v0.13.0. Deprecated, replaced with entry()::or_insert_with_if()

pub fn optionally_get_with<F>(&self, key: K, init: F) -> Option<V>
where F: FnOnce() -> Option<V>,

Returns a clone of the value corresponding to the key. If the value does not exist, evaluates the init closure, and inserts the value if Some(value) was returned. If None was returned from the closure, this method does not insert a value and returns None.

Concurrent calls on the same key

This method guarantees that concurrent calls on the same not-existing key are coalesced into one evaluation of the init closure. Only one of the calls evaluates its closure, and other calls wait for that closure to complete.

The following code snippet demonstrates this behavior:

use moka::sync::Cache;
use std::{path::Path, thread};

/// This function tries to get the file size in bytes.
fn get_file_size(thread_id: u8, path: impl AsRef<Path>) -> Option<u64> {
    println!("get_file_size() called by thread {thread_id}.");
    std::fs::metadata(path).ok().map(|m| m.len())
}

let cache = Cache::new(100);

// Spawn four threads.
let threads: Vec<_> = (0..4_u8)
    .map(|thread_id| {
        let my_cache = cache.clone();
        thread::spawn(move || {
            println!("Thread {thread_id} started.");

            // Try to insert and get the value for key1. Although all four
            // threads will call `optionally_get_with` at the same time,
            // get_file_size() must be called only once.
            let value = my_cache.optionally_get_with(
                "key1",
                || get_file_size(thread_id, "./Cargo.toml"),
            );

            // Ensure the value exists now.
            assert!(value.is_some());
            assert!(my_cache.get(&"key1").is_some());

            println!(
                "Thread {thread_id} got the value. (len: {})",
                value.unwrap()
            );
        })
    })
    .collect();

// Wait all threads to complete.
threads
    .into_iter()
    .for_each(|t| t.join().expect("Thread failed"));

Result

get_file_size() was called exactly once by thread 0.

Other threads were blocked until thread 0 inserted the value.

Thread 0 started.
Thread 1 started.
Thread 2 started.
get_file_size() called by thread 0.
Thread 3 started.
Thread 2 got the value. (len: 1466)
Thread 0 got the value. (len: 1466)
Thread 1 got the value. (len: 1466)
Thread 3 got the value. (len: 1466)

Panics

This method panics when the init closure has panicked. When it happens, only the caller whose init closure panicked will get the panic (e.g. only thread 1 in the above sample). If there are other calls in progress (e.g. thread 0, 2 and 3 above), this method will restart and resolve one of the remaining init closure.

pub fn optionally_get_with_by_ref<F, Q>(&self, key: &Q, init: F) -> Option<V>
where F: FnOnce() -> Option<V>, Q: Equivalent<K> + ToOwned<Owned = K> + Hash + ?Sized,

Similar to optionally_get_with, but instead of passing an owned key, you can pass a reference to the key. If the key does not exist in the cache, the key will be cloned to create new entry in the cache.

pub fn try_get_with<F, E>(&self, key: K, init: F) -> Result<V, Arc<E>>
where F: FnOnce() -> Result<V, E>, E: Send + Sync + 'static,

Returns a clone of the value corresponding to the key. If the value does not exist, evaluates the init closure, and inserts the value if Ok(value) was returned. If Err(_) was returned from the closure, this method does not insert a value and returns the Err wrapped by std::sync::Arc.

Concurrent calls on the same key

This method guarantees that concurrent calls on the same not-existing key are coalesced into one evaluation of the init closure (as long as these closures return the same error type). Only one of the calls evaluates its closure, and other calls wait for that closure to complete.

The following code snippet demonstrates this behavior:

use moka::sync::Cache;
use std::{path::Path, thread};

/// This function tries to get the file size in bytes.
fn get_file_size(thread_id: u8, path: impl AsRef<Path>) -> Result<u64, std::io::Error> {
    println!("get_file_size() called by thread {thread_id}.");
    Ok(std::fs::metadata(path)?.len())
}

let cache = Cache::new(100);

// Spawn four threads.
let threads: Vec<_> = (0..4_u8)
    .map(|thread_id| {
        let my_cache = cache.clone();
        thread::spawn(move || {
            println!("Thread {thread_id} started.");

            // Try to insert and get the value for key1. Although all four
            // threads will call `try_get_with` at the same time,
            // get_file_size() must be called only once.
            let value = my_cache.try_get_with(
                "key1",
                || get_file_size(thread_id, "./Cargo.toml"),
            );

            // Ensure the value exists now.
            assert!(value.is_ok());
            assert!(my_cache.get(&"key1").is_some());

            println!(
                "Thread {thread_id} got the value. (len: {})",
                value.unwrap()
            );
        })
    })
    .collect();

// Wait all threads to complete.
threads
    .into_iter()
    .for_each(|t| t.join().expect("Thread failed"));

Result

get_file_size() was called exactly once by thread 1.

Other threads were blocked until thread 1 inserted the value.

Thread 1 started.
Thread 2 started.
get_file_size() called by thread 1.
Thread 3 started.
Thread 0 started.
Thread 2 got the value. (len: 1466)
Thread 0 got the value. (len: 1466)
Thread 1 got the value. (len: 1466)
Thread 3 got the value. (len: 1466)

Panics

This method panics when the init closure has panicked. When it happens, only the caller whose init closure panicked will get the panic (e.g. only thread 1 in the above sample). If there are other calls in progress (e.g. thread 0, 2 and 3 above), this method will restart and resolve one of the remaining init closure.

pub fn try_get_with_by_ref<F, E, Q>( &self, key: &Q, init: F, ) -> Result<V, Arc<E>>
where F: FnOnce() -> Result<V, E>, E: Send + Sync + 'static, Q: Equivalent<K> + ToOwned<Owned = K> + Hash + ?Sized,

Similar to try_get_with, but instead of passing an owned key, you can pass a reference to the key. If the key does not exist in the cache, the key will be cloned to create new entry in the cache.

pub fn insert(&self, key: K, value: V)

Inserts a key-value pair into the cache.

If the cache has this key present, the value is updated.

pub fn invalidate<Q>(&self, key: &Q)
where Q: Equivalent<K> + Hash + ?Sized,

Discards any cached value for the key.

If you need to get a the value that has been discarded, use the remove method instead.

The key may be any borrowed form of the cache’s key type, but Hash and Eq on the borrowed form must match those for the key type.

pub fn remove<Q>(&self, key: &Q) -> Option<V>
where Q: Equivalent<K> + Hash + ?Sized,

Discards any cached value for the key and returns a clone of the value.

If you do not need to get the value that has been discarded, use the invalidate method instead.

The key may be any borrowed form of the cache’s key type, but Hash and Eq on the borrowed form must match those for the key type.

pub fn invalidate_all(&self)

Discards all cached values.

This method returns immediately by just setting the current time as the invalidation time. get and other retrieval methods are guaranteed not to return the entries inserted before or at the invalidation time.

The actual removal of the invalidated entries is done as a maintenance task driven by a user thread. For more details, see the Maintenance Tasks section in the crate level documentation.

Like the invalidate method, this method does not clear the historic popularity estimator of keys so that it retains the client activities of trying to retrieve an item.

pub fn invalidate_entries_if<F>( &self, predicate: F, ) -> Result<PredicateId, PredicateError>
where F: Fn(&K, &V) -> bool + Send + Sync + 'static,

Discards cached values that satisfy a predicate.

invalidate_entries_if takes a closure that returns true or false. The closure is called against each cached entry inserted before or at the time when this method was called. If the closure returns true that entry will be evicted from the cache.

This method returns immediately by not actually removing the invalidated entries. Instead, it just sets the predicate to the cache with the time when this method was called. The actual removal of the invalidated entries is done as a maintenance task driven by a user thread. For more details, see the Maintenance Tasks section in the crate level documentation.

Also the get and other retrieval methods will apply the closure to a cached entry to determine if it should have been invalidated. Therefore, it is guaranteed that these methods must not return invalidated values.

Note that you must call CacheBuilder::support_invalidation_closures at the cache creation time as the cache needs to maintain additional internal data structures to support this method. Otherwise, calling this method will fail with a PredicateError::InvalidationClosuresDisabled.

Like the invalidate method, this method does not clear the historic popularity estimator of keys so that it retains the client activities of trying to retrieve an item.

pub fn iter(&self) -> Iter<'_, K, V>

Creates an iterator visiting all key-value pairs in arbitrary order. The iterator element type is (Arc<K>, V), where V is a clone of a stored value.

Iterators do not block concurrent reads and writes on the cache. An entry can be inserted to, invalidated or evicted from a cache while iterators are alive on the same cache.

Unlike the get method, visiting entries via an iterator do not update the historic popularity estimator or reset idle timers for keys.

Guarantees

In order to allow concurrent access to the cache, iterator’s next method does not guarantee the following:

It does not guarantee to return a key-value pair (an entry) if its key has been inserted to the cache after the iterator was created.

Such an entry may or may not be returned depending on key’s hash and timing.

and the next method guarantees the followings:

It guarantees not to return the same entry more than once.

It guarantees not to return an entry if it has been removed from the cache after the iterator was created.

Note: An entry can be removed by following reasons:

Manually invalidated.

Expired (e.g. time-to-live).

Evicted as the cache capacity exceeded.

Examples

use moka::sync::Cache;

let cache = Cache::new(100);
cache.insert("Julia", 14);

let mut iter = cache.iter();
let (k, v) = iter.next().unwrap(); // (Arc<K>, V)
assert_eq!(*k, "Julia");
assert_eq!(v, 14);

assert!(iter.next().is_none());

pub fn run_pending_tasks(&self)

Performs any pending maintenance operations needed by the cache.